- Konstruktor für BasicService hinzugefügt, damit man nicht immer die Klasse übergeben muss - Eine Post Methode in der UsersResource sollte eigentlich eine GET Methode sein, ist umgebaut - queryQuestionsForLearningStateAndUser (get marked questions) hinzugefügt, kann man jetzt aufrufen: http://127.0.0.1:8050/user/state/0/3

+++ src/main/java/lernapp/model/Course.java
@@ -10,7 +10,7 @@ public class Course {
 
     @Id @GeneratedValue(strategy=GenerationType.IDENTITY)
     @Column
-    private int courseID;
+    private long courseID;
 
     @Column
     private String courseName;
@@ -29,7 +29,7 @@ public class Course {
     public void setCourseID(int courseID) {
         this.courseID = courseID;
     }
-    public int getCourseID() {
+    public long getCourseID() {
         return courseID;
     }
     public void setCourseName(String courseName) {
